Statlab_IA_Project/analysis-service/requirements.txt
fastapi==0.109
uvicorn==0.23
pandas==2.2
numpy==1.26
scipy==1.11
statsmodels==0.14
scikit-learn==1.4
xgboost==2.0
kafka-python==2.1.0
minio==7.1.0
joblib==1.3
requests==2.31.0
